move test back to client_test
